fix(faces): Docstrings suppressed same as comments
I noticed this difference comparing a Python file with this theme in Emacs and under the
Solarized theme that comes with VSCode and I preferred the latter.

I can understand the logic, docstrings aren't actually the same as literal strings from
the perspective of executing the code. Changing a literal string can display a typo to
the user or break code while without abusing docstrings changing a docstring should
never appear to the user and should never affect execution.

OTOH, in practice docstrings *do* end up displayed to users, even if just in the
published docs, and *can* affect execution. Most importantly, many languages parse
docstrings the same or similarly to literal strings so displaying them the same as
comments takes the reader further away from actual behavior. For example, a comment
immediately following a docstring bleeds into the docstring when reading making it
harder to distinguish the two.